api/requirements.txt
appdirs==1.4.4
asgiref==3.5.0
asttokens==2.0.5
attrs==21.4.0
autopep8==1.6.0
backcall==0.2.0
beautifulsoup4==4.10.0
biocommons.seqrepo==0.6.5
bioinfo-toolset @ git+https://github.com/dameyerdave/bioinfo-toolset.git@master
biopython==1.79
bioutils==0.5.5
bs4==0.0.1
canonicaljson==1.6.0
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
colored==1.4.3
coloredlogs==15.0.1
configparser==5.2.0
cryptography==36.0.1
cssselect==1.1.0
decorator==5.1.1
Django==4.0.3
django-computed-property==0.3.0
django-cors-headers==3.11.0
django-filter==21.1
django-mkdocs @ git+https://github.com/dameyerdave/django-mkdocs.git@2d55f704e6be1511eec7d6fd46e665ed86bc460d
django-phonenumber-field[phonenumbers]==6.1.0
djangorestframework==3.13.1
docker==5.0.3
drf-jwt==1.19.2
drf-schema-adapter==2.1.1
django-rest-multiple-models==2.1.3
django-simple-history==3.0.0
executing==0.8.3
fake-useragent==0.1.11
flake8==4.0.1
friendlylog==1.0.2
ga4gh.vrs==0.7.4
ghp-import==2.0.2
hgvs==1.5.2
humanfriendly==10.0
idna==3.3
importlib-metadata==4.11.2
inflection==0.5.1
Inflector==3.0.1
ipython==8.1.1
jedi==0.18.1
Jinja2==3.0.3
jsonschema==3.2.0
liftover==1.1.11
lxml==4.8.0
Markdown==3.3.6
MarkupSafe==2.1.0
matplotlib-inline==0.1.3
mccabe==0.6.1
mergedeep==1.3.4
mkdocs==1.2.3
mkdocs-include-markdown-plugin==3.3.0
mkdocs-material==8.2.4
mkdocs-material-extensions==1.0.3
numpy==1.22.3
packaging==21.3
parse==1.19.0
Parsley==1.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.0.1
prompt-toolkit==3.0.28
psycopg2==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pycodestyle==2.8.0
pycparser==2.21
pyee==8.2.2
pyflakes==2.4.0
Pygments==2.11.2
PyJWT==1.7.1
pymdown-extensions==9.2
pyparsing==3.0.7
pyppeteer==1.0.2
pyquery==1.4.3
pyrsistent==0.18.1
pysam==0.18.0
python-dateutil==2.8.2
python-jsonschema-objects==0.4.1
pytz==2021.3
PyYAML==6.0
pyyaml_env_tag==0.1
requests==2.27.1
requests-html==0.10.0
simplejson==3.17.6
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
stack-data==0.2.0
tabulate==0.8.9
termcolor==1.1.0
toml==0.10.2
tqdm==4.63.0
traitlets==5.1.1
urllib3==1.26.8
w3lib==1.22.0
watchdog==2.1.6
wcwidth==0.2.5
websocket-client==1.3.1
websockets==10.2
yoyo-migrations==7.3.2
zipp==3.7.0
